Move Json report Source Files to new namespace
