+++ src/Stryker.Core/Stryker.Core.UnitTest/MutantFilters/DashboardMutantFilterTests.cs
@@ -12,6 +12,7 @@ using Stryker.Core.Options;
 using Stryker.Core.ProjectComponents;
 using Stryker.Core.Reporters;
 using Stryker.Core.Reporters.Json;
+using Stryker.Core.Reporters.Json.SourceFiles;
 using Stryker.Core.UnitTest.Reporters;
 using Xunit;
