+++ src/Stryker.Core/Stryker.Core/Baseline/Utils/IBaselineMutantHelper.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using Stryker.Core.Mutants;
-using Stryker.Core.Reporters.Json;
+using Stryker.Core.Reporters.Json.SourceFiles;
 
 namespace Stryker.Core.Baseline.Utils
 {
