+++ src/Stryker.Core/Stryker.Core/MutantFilters/DashboardMutantFilter.cs
@@ -11,6 +11,7 @@ using Stryker.Core.Mutants;
 using Stryker.Core.Options;
 using Stryker.Core.ProjectComponents;
 using Stryker.Core.Reporters.Json;
+using Stryker.Core.Reporters.Json.SourceFiles;
 
 namespace Stryker.Core.MutantFilters
 {
