+++ src/Stryker.Core/Stryker.Core/Reporters/Json/JsonMutantLocation.cs
