+++ src/Stryker.Core/Stryker.Core/Reporters/Json/JsonMutantPosition.cs
