+++ src/Stryker.Core/Stryker.Core/Reporters/Json/JsonReport.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Stryker.Core.Options;
 using Stryker.Core.ProjectComponents;
+using Stryker.Core.Reporters.Json.SourceFiles;
 
 namespace Stryker.Core.Reporters.Json
 {
