+++ src/Stryker.Core/Stryker.Core/Reporters/Json/JsonReportSerialization.cs
