+++ src/Stryker.Core/Stryker.Core/Reporters/Json/JsonReporter.cs
