+++ src/Stryker.Core/Stryker.Core/Reporters/Json/SourceFiles/JsonMutant.cs
@@ -1,5 +1,5 @@
-﻿
-namespace Stryker.Core.Reporters.Json
+
+namespace Stryker.Core.Reporters.Json.SourceFiles
 {
     public class JsonMutant
     {
